Struct interpreter::DefaultRuntime

pub struct DefaultRuntime<'db, 'msg, 'st, 'sys, 'r, BS, SYS> { /* fields omitted */ }

Implementation of the Runtime trait.

Implementations

impl<'db, 'msg, 'st, 'sys, 'r, BS, SYS> DefaultRuntime<'db, 'msg, 'st, 'sys, 'r, BS, SYS> where
BS: BlockStore,
SYS: Syscalls,

pub fn new(
state: &'st mut StateTree<'db, BS>,
store: &'db BS,
syscalls: &'sys SYS,
gas_used: i64,
message: &'msg UnsignedMessage,
epoch: ChainEpoch,
origin: Address,
origin_nonce: u64,
num_actors_created: u64,
rand: &'r ChainRand
) -> Self

Constructs a new Runtime

pub fn charge_gas(&mut self, to_use: i64) -> Result<(), ActorError>

Adds to amount of used

Will borrow gas tracker RefCell, do not call if any reference to this exists

pub fn gas_used(&self) -> i64

Returns gas used by runtime

Will borrow gas tracker RefCell, do not call if a mutable reference exists

pub fn price_list(&self) -> PriceList

Returns the price list for gas charges within the runtime