Crate interpreter

Structs

ApplyRet
Apply message return data.

BlockMessages
Contains all messages to process through the VM as well as miner information for block rewards.

DefaultRuntime
Implementation of the Runtime trait.

GasCharge
Single gas charge in the VM. Contains information about what gas was for, as well as the amount of gas needed for computation and storage respectively.

PriceList
Provides prices for operations in the VM

VM
Interpreter which handles execution of state transitioning messages and returns receipts from the vm execution.

Traits

CircSupplyCalc
Allows generation of the current circulating supply given some context.

LookbackStateGetter
Trait to allow VM to retrieve state at an old epoch.

Rand
Randomness provider trait

Functions

price_list_by_epoch
Returns gas price list by Epoch for gas consumption.

resolve_to_key_addr
returns the public key type of address (BLS/SECP256K1) of an account actor identified by addr.