Struct interpreter::DefaultRuntime

pub struct DefaultRuntime<'db, 'vm, BS, R, C, LB, V, P = DefaultNetworkParams> { /* fields omitted */ }

Implementation of the Runtime trait.

Implementations

impl<'db, 'vm, BS, R, C, LB, V, P> DefaultRuntime<'db, 'vm, BS, R, C, LB, V, P> where
BS: BlockStore,
V: ProofVerifier,
P: NetworkParams,
R: Rand,
C: CircSupplyCalc,
LB: LookbackStateGetter<'db, BS>,

pub fn new(
version: NetworkVersion,
state: &'vm mut StateTree<'db, BS>,
store: &'db BS,
gas_used: i64,
base_fee: TokenAmount,
message: &UnsignedMessage,
epoch: ChainEpoch,
origin: Address,
origin_nonce: u64,
num_actors_created: u64,
depth: u64,
rand: &'vm R,
registered_actors: &'vm HashSet<Cid>,
circ_supply_calc: &'vm C,
lb_state: &'vm LB
) -> Result<Self, ActorError>

Constructs a new Runtime

pub fn charge_gas(&mut self, gas: GasCharge) -> Result<(), ActorError>

Adds to amount of used.

Will borrow gas tracker RefCell, do not call if any reference to this exists

pub fn gas_used(&self) -> i64

Returns gas used by runtime.

Will borrow gas tracker RefCell, do not call if a mutable reference exists

pub fn price_list(&self) -> &PriceList

Returns the price list for gas charges within the runtime.

pub fn send(
&mut self,
msg: &UnsignedMessage,
gas_cost: Option<GasCharge>
) -> Result<Serialized, ActorError>

Shared logic between the DefaultRuntime and the Interpreter. It invokes methods on different Actors based on the Message. This function is somewhat equivalent to the go implementation’s vm send.

pub fn try_create_account_actor(
&mut self,
addr: &Address
) -> Result<(ActorState, Address), ActorError>

creates account actors from only BLS/SECP256K1 addresses.